chore: move source file of vite-config package inside lib folder
