+++ packages/vite-config/lib/withPageConfig.mjs
